Update table_info call to get back table data for test SQLite DB
The dump_current_and_savepoint function in shredder/utils.pl relies
on the table_info call from DBD::SQLite to return all of the tables in
the test RT database. The call was using empty strings as the first
two parameters, but instead of returning all tables, it returned none.

The first parameter is catalog which isn't supported in SQLite,
and the second, schema, seemed to be performing a LIKE and matching
nothing.

Based on the examples in the DBD::SQLite tests, passing undef
for the first two parameters and '%' for the third parameter, table,
returns all tables.

This change returns data for the test comparison and in
doing so reveals a failing test in t/shredder/01ticket.t