factor out ExtractTicketId function on top of ParseTicketId
it would be very handy for any extension that wants
to implement different algorithms for attaching
messages to tickets, for example by reference
headers or by ids from external systems.

First idea was to pass MIME::Entity right into
ParseTicketId function as second argument, but
this variant is more flexible. With this change
ExtractTicketId can be overriden to change where
id searched, but still have access to RT's
ParseTicketId that just finds id in any string.