Standard error pages for all reasons we abort external auth
Instead of a blank page, show brief, friendlier error messages with
information about why auth failed, contact information for the local
admin, and an internal login link when fallback is on.  Rendering each
error as an individual component allows easier site-specific overrides
for customizing messages to the external authentication system in use.

Behaviour under WebFallbackToInternalAuth is changed from showing login
pages with no message on failure to showing an error message with a way
to login.  Below is a summary of the new behaviour for each reason we
abort and force logout during WebExternalAuth checking.  All error pages
are 403s and provide a login link if Fallback is ON.

User doesn't exist and auto-create is OFF
    Fallback  ON: Error page, since if they were supposed to login it would
                  likely be linked to remote user

    Fallback OFF: Error page

Fail to auto-create a user (iff WebExternalAuto is ON)
    Fallback  ON: Error page with "couldn't create user message", since they're
                  unlikely to be able to do anything about it.

    Fallback OFF: Error page with "couldn't create user" message

User logged in but deauthed by external system (iff WebExternalAuthContinuous is ON)
    Fallback  ON: Error page, suggesting refresh. Next page hit gets them
                  logged in again if it was a blip or the login form if it wasn't.

    Fallback OFF: Same error page as above, except they'll get the generic
                  error page below if it wasn't a blip when they refresh.

No remote user
    Fallback  ON: Login page
    Fallback OFF: Error page, generic